requirements.txt
python-dotenv
mysql-connector-python
pandas
Faker
numpy